changes to config file
