+++ network/test/test_orca_setup_config.yaml
@@ -2,3 +2,5 @@ sonic_ips:
   - 10.10.229.112
 onie_ips:
   - 10.10.229.123
+sonic_image_4.4.0_url: http://10.10.128.249/sonic/release/4.4.0/sonic-broadcom-enterprise-advanced.bin
+socic_image_4.2.0_url: http://10.10.128.249/sonic/release/4.2.0/sonic-broadcom-enterprise-advanced.bin
